requirements.txt
asgiref==3.7.2
Django==4.0
django-ckeditor==6.7.1
django-filter==23.5
django-js-asset==2.2.0
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
